requirements.txt
Flask>=2.2.2
Flask-WTF>=1.1.1
python-dotenv>=0.21.0
Flask-Caching>=1.10.1
gunicorn>=20.1.0
